CLOUD-65216 echo back Ambari and HDP version in stack response
