+++ core-api/src/main/java/com/sequenceiq/cloudbreak/doc/ModelDescriptions.java
@@ -51,7 +51,6 @@ public class ModelDescriptions {
 
     public static class ImageModelDescription {
         public static final String IMAGE_NAME = "name of the image";
-        public static final String HDPVERSION = "hdp version of image";
     }
 
     public static class CloudbreakDetailsModelDescription {
