+++ core/src/main/java/com/sequenceiq/cloudbreak/converter/StackToJsonConverter.java
@@ -19,7 +19,9 @@ import com.sequenceiq.cloudbreak.api.model.ImageJson;
 import com.sequenceiq.cloudbreak.api.model.InstanceGroupJson;
 import com.sequenceiq.cloudbreak.api.model.OrchestratorResponse;
 import com.sequenceiq.cloudbreak.api.model.StackResponse;
+import com.sequenceiq.cloudbreak.cloud.model.AmbariRepo;
 import com.sequenceiq.cloudbreak.cloud.model.CloudbreakDetails;
+import com.sequenceiq.cloudbreak.cloud.model.HDPRepo;
 import com.sequenceiq.cloudbreak.cloud.model.Image;
 import com.sequenceiq.cloudbreak.core.CloudbreakImageNotFoundException;
 import com.sequenceiq.cloudbreak.domain.InstanceGroup;
@@ -49,6 +51,7 @@ public class StackToJsonConverter extends AbstractConversionServiceAwareConverte
         } catch (CloudbreakImageNotFoundException exc) {
             LOGGER.info(exc.getMessage());
         }
+
         stackJson.setName(source.getName());
         stackJson.setOwner(source.getOwner());
         stackJson.setAccount(source.getAccount());
@@ -90,14 +93,29 @@ public class StackToJsonConverter extends AbstractConversionServiceAwareConverte
         }
         stackJson.setCreated(source.getCreated());
         stackJson.setGatewayPort(source.getGatewayPort());
+        convertComponentConfig(stackJson, source.getId());
+        return stackJson;
+    }
+
+    private StackResponse convertComponentConfig(StackResponse stackJson, Long stackId) {
         try {
-            CloudbreakDetails cloudbreakDetails = componentConfigProvider.getCloudbreakDetails(source.getId());
+            AmbariRepo ambariRepo = componentConfigProvider.getAmbariRepo(stackId);
+            if (ambariRepo != null) {
+                stackJson.setAmbariVersion(ambariRepo.getVersion());
+            }
+            HDPRepo hdpRepo = componentConfigProvider.getHDPRepo(stackId);
+            if (hdpRepo != null) {
+                stackJson.setHdpVersion(hdpRepo.getHdpVersion());
+            }
+
+            CloudbreakDetails cloudbreakDetails = componentConfigProvider.getCloudbreakDetails(stackId);
             if (cloudbreakDetails != null) {
                 stackJson.setCloudbreakDetails(getConversionService().convert(cloudbreakDetails, CloudbreakDetailsJson.class));
             }
         } catch (Exception e) {
-            LOGGER.error("Cloudbreak details could not be added to Stack response.", e);
+            LOGGER.error("Failed to convert dynamic component.", e);
         }
+
         return stackJson;
     }
